+++ test/UiAdminBase/Action/UiActionParameterTest.php
@@ -5,12 +5,16 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Ui\Admin\Base\Test\Action;
 
 use Heptacom\HeptaConnect\Core\Test\Fixture\FooBarEmitter;
+use Heptacom\HeptaConnect\Core\Test\Fixture\FooBarPortalExtension;
 use Heptacom\HeptaConnect\Dataset\Base\Contract\AttachmentAwareInterface;
 use Heptacom\HeptaConnect\Portal\Base\FlowComponent\CodeOrigin;
+use Heptacom\HeptaConnect\Storage\Base\PreviewPortalNodeKey;
 use Heptacom\HeptaConnect\Storage\Base\Test\Fixture\FirstEntity;
 use Heptacom\HeptaConnect\Storage\Base\Test\Fixture\Portal;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Action\Portal\PortalEntityList\PortalEntityListCriteria;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Action\Portal\PortalEntityList\PortalEntityListResult;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionBrowse\PortalNodeExtensionBrowseCriteria;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionBrowse\PortalNodeExtensionBrowseResult;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -19,8 +23,11 @@ use PHPUnit\Framework\TestCase;
  * @covers \Heptacom\HeptaConnect\Dataset\Base\Support\AbstractCollection
  * @covers \Heptacom\HeptaConnect\Dataset\Base\Support\AbstractObjectCollection
  * @covers \Heptacom\HeptaConnect\Portal\Base\FlowComponent\CodeOrigin
+ * @covers \Heptacom\HeptaConnect\Storage\Base\PreviewPortalNodeKey
  * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Action\Portal\PortalEntityList\PortalEntityListCriteria
  * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Action\Portal\PortalEntityList\PortalEntityListResult
+ * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionBrowse\PortalNodeExtensionBrowseCriteria
+ * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionBrowse\PortalNodeExtensionBrowseResult
  */
 final class UiActionParameterTest extends TestCase
 {
@@ -42,8 +49,12 @@ final class UiActionParameterTest extends TestCase
         $portalClass = Portal::class;
         $entityType = FirstEntity::class;
         $codeOrigin = new CodeOrigin(__FILE__, 0, 1);
+        $portalNodeKey = new PreviewPortalNodeKey($portalClass);
+        $portalExtensionClass = FooBarPortalExtension::class;
 
         yield new PortalEntityListCriteria($portalClass);
         yield new PortalEntityListResult($codeOrigin, $entityType, FooBarEmitter::class);
+        yield new PortalNodeExtensionBrowseCriteria($portalNodeKey);
+        yield new PortalNodeExtensionBrowseResult($portalNodeKey, true, $portalExtensionClass);
     }
 }